Create UpdateCOE.PS1
This basic PowerShell script will update an existing COE installation from a PowerShell command prompt. IMHO, much easier than clicking through the maker portal and list of solutions. 

I was able to test this with the 12/7/2022 release. This script presumes static solutions names (less the appended version). The side effect of this presumption is that new solutions will not be installed unless they are added to the script. New solutions requiring new connections or custom connectors, etc. will not work for initial deployment. This script is meant to update existing installation.